requirements.txt
flask
torch
gtts
openai-whisper

